Config class definition

This class is used to store the configuration of any process used with MagellanNTK It contains a validity function to ensure that the format is correct.

Validity: * The first step must be called 'Description', it is a mandatory step. Thus, the first item of the mandatory vector is TRUE. To be continued...

## Initialization ### Generic process

A generic process * Generic pipeline : xxxx * Description pipeline: This case is for a process called 'Description' which is the first process module of a pipeline

Wrapper function to the constructor of the class

Usage

Config(
  fullname = "",
  mode = "",
  steps = NULL,
  mandatory = NULL,
  steps.source.file = NULL
)

# S4 method for class 'Config'
initialize(.Object, fullname, mode, steps, mandatory, steps.source.file)

Config(
  fullname = "",
  mode = "",
  steps = NULL,
  mandatory = NULL,
  steps.source.file = NULL
)

Arguments

fullname

The complete name of the pipeline or the process (in this case, it is the concatenation of the name of the pipeline and the name of the process itself, separated by '_')

mode

A `character()` which indicates whether the current module is used as a 'process' nor a 'pipeline.' Default value is NULL

steps

A vector containing the names of steps in a process or a pipeline.

mandatory

A vector of boolean where each element indicates whether the corresponding steps is mandatory or not. It has the same length of the vector steps.

steps.source.file

A `vector` of which each item is the source code of the corresponding step.

.Object

xxx

Value

NA

Slots

fullname

The name of the process which is the concatenation of the name of the pipeline and the name of the process itself, separated by '_'.

name

The name of the process nor pipeline.

parent

The name of the pipeline/process which owns this instance

mode

A `character()` which indicates if the configuration is about a whole process nor a process of a pipeline.

steps

A `vector` of `character()` which contains the primary steps of the pipeline. Two steps will be automatically inserted in this vector: 'Description 'in the first position and 'Save' at the end.

mandatory

A `vector` of `boolean()` in which each item is the necessary code for the GUI of a step. The size of this vector is the number of steps

ll.UI

A `vector` of Shiny source code. Each item is the necessary code for the GUI of a step. The size of this vector is the number of steps

steps.source.file

A `vector` of which each item is the source code of the corresponding step.

steps

A `vector` of `character()` which contains the primary steps of the pipeline. Two steps will be automatically inserted in this vector: 'Description 'in the first position and 'Save' at the end.

mandatory

A `vector` of `boolean()` in which each item is the necessary code for the GUI of a step. The size of this vector is the number of steps

Examples

# Example of a generic process
generic.proc <- Config(
    fullname = "PipelineDemo_Process1",
    mode = "process",
    steps = c("Step 1", "Step 2"),
    mandatory = c(TRUE, FALSE)
)

# Example of a generic pipeline
generic.pipe <- Config(
    fullname = "Pipe1_PipelineDemo",
    mode = "pipeline",
    steps = c("Process 1", "Process 2"),
    mandatory = c(TRUE, FALSE)
)

# Example of a root pipeline (process has no parent)
root.pipe <- Config(
    mode = "pipeline",
    fullname = "PipelineDemo",
    steps = c("Process1", "Process2 bis", "Process3"),
    mandatory = c(FALSE, FALSE, TRUE)
)

# Example of a description module (process has no steps)
description.process <- Config(
    fullname = "PipelineDemo_Description",
    mode = "process",
    steps = "",
    mandatory = ""
)

generic.proc
#> 	 ------- Config -------
#> 	fullname: PipelineDemo_Process1
#> 	name: Process1
#> 	parent: PipelineDemo
#> 	mode: process
#> 	names(steps): Description Step1 Step2 Save
#> 	steps: Description Step 1 Step 2 Save
#> 	mandatory: TRUE TRUE FALSE TRUE
#> 	names(ll.UI): 
#> 	ll.UI: 
#> 
generic.pipe
#> 	 ------- Config -------
#> 	fullname: Pipe1_PipelineDemo
#> 	name: PipelineDemo
#> 	parent: Pipe1
#> 	mode: pipeline
#> 	names(steps): Description Process1 Process2 Save
#> 	steps: Description Process 1 Process 2 Save
#> 	mandatory: TRUE TRUE FALSE TRUE
#> 	names(ll.UI): 
#> 	ll.UI: 
#> 
root.pipe
#> 	 ------- Config -------
#> 	fullname: PipelineDemo
#> 	name: PipelineDemo
#> 	parent: 
#> 	mode: pipeline
#> 	names(steps): Description Process1 Process2bis Process3 Save
#> 	steps: Description Process1 Process2 bis Process3 Save
#> 	mandatory: TRUE FALSE FALSE TRUE TRUE
#> 	names(ll.UI): 
#> 	ll.UI: 
#> 
description.process
#> 	 ------- Config -------
#> 	fullname: PipelineDemo_Description
#> 	name: Description
#> 	parent: PipelineDemo
#> 	mode: process
#> 	names(steps): 
#> 	steps: 
#> 	mandatory: 
#> 	names(ll.UI): 
#> 	ll.UI: 
#>